+++ CHANGES.md
@@ -2,6 +2,7 @@
 - [FIXED] Add URL to `CouchDbException` exception message (where applicable) for easier debugging.
 - [NEW] Add additional method to `GET` standalone attachments.
 - [FIXED] Issue with "+" (plus) not being regarded as a reserved character in URI path components.
+- [IMPROVED] Faster shutdown when using the optional OkHttp client.
 - [FIXED] Issue with double encoding of restricted URL characters in credentials when using
    `ClientBuilder.url()`.
 - [NEW] Added `bluemix` method to the client builder allowing service credentials to be passed using